projects/learning-airflow/requirements.txt
# Add Python packages here, one per line. Pin versions where possible.
apache-airflow-providers-snowflake==6.8.1
apache-airflow-providers-http==5.6.2